requirements.txt
hbutils>=0.10.0
hfutils>=0.3.4
huggingface_hub
tqdm
requests
click>=7
natsort
pillow
dghs-imgutils>=0.4.8
numpy<2
pandas
pyyaml>=6
pyarrow